pyproject.toml
[tool.poetry]
name = "backbase"
version = "0.1.0"
description = ""
authors = ["Premsai Kududula <premsaik98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
requests = "^2.32.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-environ = "^0.11.2"
django-constance = "^4.1.2"
django-extensions = "^3.2.3"
ipython = "^8.28.0"
django-redis = "^5.4.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
